Merge `matches` polyfill from legacy into main file
This eases the users' burden of deciding which implementation to load,
simplifies maintenance by not having two separate implementation files
which are mostly duplicated, and makes the compatibility matrix easier
to read.

The downside is that it increases the size of the main polyfill
slightly, but that shouldn't be a concern because the total size of the
polyfill is negligible anyway.